backend/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
boto3==1.28.84
botocore==1.31.84
click==8.1.7
decouple==0.0.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg==3.1.12
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
